% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{engine_K3d}
\alias{engine_K3d}
\title{Engine for computing K-function}
\usage{
engine_K3d(r, s, x_vec, y_vec, z_vec, dists_sph, Dmat)
}
\arguments{
\item{r}{Vector of values for the argument r at which K(r, s) should be evaluated.}

\item{s}{Vector of values for the argument s at which K(r, s) should be evaluated.}

\item{x_vec}{Vector of x coordinates.}

\item{y_vec}{Vector of y coordinates.}

\item{z_vec}{Vector of z coordinates.}

\item{dists_sph}{Matrix of distances between points on S^2.}

\item{Dmat}{Matrix of all possible values to sum over to.
The values of the K-function before multiplying by the indicator function.
Constants may be multiplied on the resulting matrix.}
}
\description{
Engine for calculating K-function of point processes on the product space of R^3 and S^2.
}
