% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{engine_k3s}
\alias{engine_k3s}
\title{Engine for computing K-function}
\usage{
engine_k3s(r, s, dists_3d, dists_sph, Dmat)
}
\arguments{
\item{r}{Vector of values for the argument r at which K(r, s) should be evaluated.}

\item{s}{Vector of values for the argument s at which K(r, s) should be evaluated.}

\item{dists_3d}{Matrix of distances between points in R^3.}

\item{dists_sph}{Matrix of distances between points on S^2.}

\item{Dmat}{Matrix of all possible values to sum over to.
The values of the K-function before multiplying by the indicator function.
Constants may be multiplied on the resulting matrix.}
}
\description{
Engine for calculating K-function of point processes on the product space of R^3 and S^2.
}
