% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K3dsph.R
\name{K3dsph}
\alias{K3dsph}
\title{K-function}
\usage{
K3dsph(X, Y, r = NULL, s = NULL, rmax = NULL, smax = NULL,
  nrval = 128, nsval = nrval, intenssX = NULL, intenssY = NULL, parmsX,
  parmsY)
}
\arguments{
\item{X}{Point pattern of class \code{\link{pp3}}.}

\item{Y}{Point pattern of class \code{\link{pps}}.}

\item{r}{Optional. Vector of values for the argument r at which K(r, s) should be evaluated.}

\item{s}{Optional. Vector of values for the argument s at which K(r, s) should be evaluated.}

\item{rmax}{Optional. Maximum desired value of the argument r.}

\item{smax}{Optional. Maximum desired value of the argument s.}

\item{nrval}{Optional. Number of values of r for which K(r, s) will be estimated.}

\item{nsval}{Optional. Number of values of s for which K(r, s) will be estimated.}

\item{intenssX}{Optional. Values of the estimated intensity function in 3d space.
Either a \code{NULL}, vector, matrix, or function.
If \code{NULL} the function will return the space-sphere K-function under the assumption of homogeneity in 3d space.
If a vector it should contain the intensities at each of the observed points.
If a matrix it should contain the product of the intensities of every pair of observed points.
If a function it should return a vector containing the intensities at each of the observed points.}

\item{intenssY}{Optional. Values of the estimated intensity function on the unitsphere.
Either a \code{NULL}, vector, matrix, or function.
If \code{NULL} the function will return the space-sphere K-function under the assumption of homogeneity on the unitsphere.
If a vector it should contain the intensities at each of the observed points.
If a matrix it should contain the product of the intensities of every pair of observed points.
If a function it should return a vector containing the intensities at each of the observed points.}

\item{parmsX}{List of additional arguments passed to \code{intenssX} if it is given as a function.}

\item{parmsY}{List of additional arguments passed to \code{intenssY} if it is given as a function.}
}
\value{
A list containing \code{r}, \code{s}, \code{theo}, and \code{K3dsph}.
\code{theo} is the theoretical space-sphere K-function under Stationary Poisson.
\code{theo} and \code{K3dsph} are matrices with \code{length(r)} rows and \code{length(s)} columns.
}
\description{
Estimates the space-sphere K-function from a point pattern with points in \code{R^3 x S^2}
in a window of arbitry shape in space and on the entire sphere.
}
\details{
If inhomogeneous it is assumed that the intensity of the space-sphere points is the product of the intensity in space and intensity on the sphere.
}
